requirements.txt
opencv-contrib-python
pandas
pillow